Update strict badge search returns (#5)
* Update gitignore for GoLand

* Return 200 empty for strict badge lookup failure

This is just to make things easier for the web app so we don't have to worry about checking for specific error code returns and just build an empty HTML on the page, which is fine.

* Update error outputs

Update the error output for the fuzzySearch function to return an accurate error message for in the case that no query parameters are provided. Also remove unnecessary if/else block.