Add startup script to copy correct pg_hba.conf file in
Previously I was attempting to copy the pg_hba.conf file into the /data directory due to some misunderstanding on how the file system was structured on my part. After realizing I needed that conf file in the /var/lib/postgresql/data directory I realized I needed to add a startup script to the container on build that would copy the contents from it's /data directory to the /var/lib/postgresql/data directory, because that dir needs to be empty on initdb, and the start up script can overwrite the automatically generated file after the initdb command is completed.