Fuck it we're going with shell scripts
Because Docker on Linux is not the same as Docker on anything else I was running into a lot of issues with file permissions, specifically around the desire to expose the Postgres server's log files back out to the host to allow for some nice fail2ban shenanigans. After some pain (see commit history) I decided that it would just be easier to just use the ./init-script.sh to spin up the docker containers, then run a quick chown with the in container postgres uid:gid, which is 70:70 and appears to be static? Hopefully? Anyway it works so what ever.

I also stopped using docker volumes and am just copying the necessary config files in via the dockerfile now as God intended.