Enable historical roster returns
Update the necessary functions to allow multiple returns for a single officer to allow the searching of historical roster records. This includes having to update the SeattleGetOfficerByBadge to return a list of officers instead of a single officer.

All this will need to be changed again but this is step 1 in getting this working in a more manageable fashion. Currently the API just returns way too many records with the historical roster data.