Historical SPD rosters + access updates (#23)
* Add reader user to postgres 

New user 'reader' created with read only permissions to database tables.

* Add external DB ports to docker-compose files

Normally we wouldn't want to expose the postgres port to the outside world but now that we are looking at allowing external users this will be required to enable it.

* Add historical SPD roster CSV to seed

* Get historical roster CSV formatted and imported properly

* Update 2_seattle.sql to use new historical roster file

* Disable remote postgres user logins

Update docker-compose to utilize new pg_hba.conf file to allow only the api service container to login to the postgres user account remotely.

Allow reader user to login from anywhere.

* Add startup script to copy correct pg_hba.conf file in

Previously I was attempting to copy the pg_hba.conf file into the /data directory due to some misunderstanding on how the file system was structured on my part. After realizing I needed that conf file in the /var/lib/postgresql/data directory I realized I needed to add a startup script to the container on build that would copy the contents from it's /data directory to the /var/lib/postgresql/data directory, because that dir needs to be empty on initdb, and the start up script can overwrite the automatically generated file after the initdb command is completed.

* Sync local build and remote pull dc files

* Enable external access to postgresql.log for fail2ban support

Also ignore the new directory.

* potential conf file permission issues fix

want to move to couch commit

* Update compose with postgresql.conf file usage

* install fail2ban

* Fuck it we're going with shell scripts

Because Docker on Linux is not the same as Docker on anything else I was running into a lot of issues with file permissions, specifically around the desire to expose the Postgres server's log files back out to the host to allow for some nice fail2ban shenanigans. After some pain (see commit history) I decided that it would just be easier to just use the ./init-script.sh to spin up the docker containers, then run a quick chown with the in container postgres uid:gid, which is 70:70 and appears to be static? Hopefully? Anyway it works so what ever.

I also stopped using docker volumes and am just copying the necessary config files in via the dockerfile now as God intended.

* Remove docker volume clean up from init-script

Not using volumes anymore so no need.

* Enable historical roster returns

Update the necessary functions to allow multiple returns for a single officer to allow the searching of historical roster records. This includes having to update the SeattleGetOfficerByBadge to return a list of officers instead of a single officer.

All this will need to be changed again but this is step 1 in getting this working in a more manageable fashion. Currently the API just returns way too many records with the historical roster data.

* Whitespace in seattle SQL functions file

* Add strict match historical and current functions

Add the is_current field to the officer object, and update the strict search functions to use direct SQL queries instead of using SQL stored procedures for simplicity.

Update name and badge searches to have historical counter parts. SQL code updated for all functions to handle historical results and either return only the most recent available roster entry per officer or return all historical roster entries unfiltered but ordered by most recent date available to oldest date available.

Plumb these new functions up through the handler and router to be exposed to the external API appropriately.

Remove all SQL functions in the sql set up file. It is now just creating the table and populating it with the data.

* Add strict match historical and current functions

Add the is_current field to the officer object, and update the strict search functions to use direct SQL queries instead of using SQL stored procedures for simplicity.

Update name and badge searches to have historical counter parts. SQL code updated for all functions to handle historical results and either return only the most recent available roster entry per officer or return all historical roster entries unfiltered but ordered by most recent date available to oldest date available.

Plumb these new functions up through the handler and router to be exposed to the external API appropriately.

Remove all SQL functions in the sql set up file. It is now just creating the table and populating it with the data.

* I have no idea what git thought it was doing there

* Fix git merge issue

Apparently the lines I deleted were a result of a git merge conflict that wasn't resolved properly.

* Fix other git merge conflict

* Simplify Docker setup

Removed the attempted piping of logs from the SQL container to the host system as it added too much complexity and wasn't that useful.

Committed a suggestion by Madison to use COPY *.conf instead of calling out the individual conf files for postgres.

* Remove strict search name ordering code

This got missed some how. New SQL queries will return the roster entires in the desired order so there is no need to do any sorting in the Go code itself. Removed a block of code that was sorting returned officers by last name instead of preserving the ordering by roster date from the SQL query.