Enable GitHub Action local testing via act library
See following repo for more information:
https://github.com/nektos/act

This update explicitly sets the GitHub token to be used to perform the checkout for the action build process in the two action files. According to the actions/checkout documentation this should not break the process when run on GitHub because it is just explicitly stating the default configuration for the checkout job. See the below link for more details:

https://github.com/actions/checkout#usage