Enable GitHub Action local testing via act library
See following repo for more information:
https://github.com/nektos/act

This update explicitly sets the GitHub token to be used to perform the checkout for the action build process in the two action files. This should be the only real change to the Actions themselves, it is unclear to me whether this will break the Actions when run on GitHubs server but I believe that the token line is just a more explicit statement of what is implied in the current configuration.