Enable local GHA testing via act repo (#27)
* Enable GitHub Action local testing via act library

See following repo for more information:
https://github.com/nektos/act

This update explicitly sets the GitHub token to be used to perform the checkout for the action build process in the two action files. This should be the only real change to the Actions themselves, it is unclear to me whether this will break the Actions when run on GitHubs server but I believe that the token line is just a more explicit statement of what is implied in the current configuration.

* Enable GitHub Action local testing via act library

See following repo for more information:
https://github.com/nektos/act

This update explicitly sets the GitHub token to be used to perform the checkout for the action build process in the two action files. According to the actions/checkout documentation this should not break the process when run on GitHub because it is just explicitly stating the default configuration for the checkout job. See the below link for more details:

https://github.com/actions/checkout#usage