Add seattleOfficer structu for internal conversion
This probably isn't the best way to do it but I'm just gonna do it this way. Basically we scan everything from the SQL query return into a struct that is a copy of the original SeattleOfficer struct, then we copy everything over and convert the datetime into a string formatted for ISO with just the date. All together not too bad.

I accidentally included a white space update from gofmt changes in this as well. My bad but I can't really extricate those changes from the substantive changes unfortunately.