Fix the datetime return formatting (#29)
* Was sorting historical returns by name for some reason

Fixed.

* Add seattleOfficer structu for internal conversion

This probably isn't the best way to do it but I'm just gonna do it this way. Basically we scan everything from the SQL query return into a struct that is a copy of the original SeattleOfficer struct, then we copy everything over and convert the datetime into a string formatted for ISO with just the date. All together not too bad.

I accidentally included a white space update from gofmt changes in this as well. My bad but I can't really extricate those changes from the substantive changes unfortunately.

* Similar update to Tacoma database module as Seattle

This is functionally exactly what I did for the seattle_db update except for Tacoma. Same process, scan SQL query results onto intermediary struct, then convert date to ISO date format and convert any nulls to empty strings.

The accidental white space changes from gofmt are much worse here than seattle_db so again sorry for that.

* Update testing fixtures to handle new date type

Since the main struct's date is now a string and not a time.Time.

Removed time as well because it's no longer needed.

* The gofmt change

All these changes are non-substantive and only a result of running gofmt to make the files prettier.

* Remove Historical Search sort

The sorting of the returned slice in the seattleGetOfficerByBadgeHistorical function is not necessary because the SQL query handles the date sort, so removing.

* Remove json descriptors on the tacomaOfficer struct

I forgot to remove the json descriptors from the tacomaOfficer struct to make it match the seattleOfficer struct.