Clean up artifacts PR (#30)
* Swap docker-compose and docker-compose-local

It was decided that it makes more sense to make the default docker-compose build from source rather than fetch from Dockerhub since that's mostly what we're doing anyway.

* Remove init-script.sh

No longer necessary as we no longer maintain volumes for the database.

* Simple script to ingest SPD rosters and format them properly for 

Pretty bare bones but it does the job.

* Fix unit_description typo

* Fix python data prep script

The script wasn't working how I remembered it working so I fixed it so it did.

* Update docker-compose-local name

This better identifies what the file is. Local wasn't the right name it because it implies that we're building the local code where as in this case we are not building but pulling from Dockerhub instead.

* Add nifty one liner to strip all whitespace from strings

Nice little one liner to make sure that any DataFrame field that is a string is stripped of excess whitespace.

* Add click interface to make script easier to deal with

Replaced use of sys.argv with click library to make the script a bit more friendly to work with.

* Remove unnecessary df.apply

Apply is bad so worked around using it.  Updated how we split names as well to fix some logic to make sure that names make it into their proper fields. May need to go back and make sure this is normalized across the entire seed dataset.

* Update historical SPD roster with June 2021 rosters