Update naming convention of roster sheet files
The existing roster sheet files were using 3 character designations to denote which policing organization they belong to, which could be a little confusing, especially with the addition of several more org rosters being added. To help make this a little clearer I updated the naming convention so we have a bit more information on the source roster and organization.