Add new police orgs to lookup (#32)
* Update naming convention of roster sheet files

The existing roster sheet files were using 3 character designations to denote which policing organization they belong to, which could be a little confusing, especially with the addition of several more org rosters being added. To help make this a little clearer I updated the naming convention so we have a bit more information on the source roster and organization.

* Add new rosters for other police orgs

Adding new rosters from a few other police orgs that have been provided to us. This is just the CSVs, subsequent commits will add the functionality to the API to populate the returns and various department information.

* Add new sql files

Add associated SQL set up files for the new police orgs:
 * Auburn
 * Lakewood
 * Olympia

Needed to update the numbers on the sql files so they get executed in the right order so had to update the read only user creation and remote access shell script file names as well.

* Add Auburn PD roster to dataset

This commit adds all the necessary files to allow users to query the API for Auburn PD officers. This allows for both strict and fuzzy searching of first names and last names, as well as strict searching by badge number.

* Add structures for Lakewood PD

This commit adds all the necessary files to allow users to query the API for Lakewood PD officers. This allows for both strict and fuzzy searching of first names and last names. No badges though because that is not included in the dataset.

* Remove historical function from auburn handler

copy/paste error

* Add structures for Olympia PD

This commit adds all the necessary files to allow users to query the API for Lakewood PD officers. This allows for both strict and fuzzy searching of first names and last names. Also add strict search for badges.