Fix roster source date
The correct date for the roster source from Tacoma was for any officer who worked for the Tacoma Police Department in the year 2019, along with the salary information that was provided. Somewhere along the line the CSV was updated to include the date the roster was added to the site, rather than the actual source date for the roster.

Because this roster data is not from a particular day, I've updated the DepartmentMetadata object to just return a simple String of 2019 instead of converting a date and using the erroneous date in the CSV column. This will be further addressed in another PR to handle the Tacoma PD data better.