Add just and mirror config for signal-scanner-bot repo
This creates a config for just command line with The following recipes:
 * default - returns list of available recipes
 * build - builds the specified docker images in the docker-compose file
 * up service="" - spins up all services specified in the docker-compose file or a specified service provided from the command line flag service
 * down - spins down all services specified in the docker-compose file
 * pull - pulls all images specified in the docker-compose file

This just config also looks for an environment variable `IS_PROD` that will utilize or not utilize a docker-compose.override file. If the env var is set just will not use the override file, if it is set it will use the override file.