Add build and test steps for pull_request action (#44)
* Add build and test steps for pull_request actions

Force use of unit testing in pull requests.

* go fmt commit

Just gonna sneak a go fmt commit in here since I've been forgetting to do that for each of my commits. Need to just make it part of my git workflows.

* Add gofmt check for GHA

This runs go fmt on the ./api directory on pull request to ensure that the code has been formatted properly to match the go standard.

* Redo go fmt changes again

Not exactly sure what's going on with these merge conflict changes but this at least reduces the changes to just formatting issues instead of having to fix the tests again.