Remove dependency on volumes
Reality is I've been playing with Kubernetes and I wanted to try using it with this app, but I didn't want to deal with persistent volumes, so here we are.

This also makes the set up a bit easier just like in general so I guess win win. I checked though and it only costs an extra 3MB for the image so... what ever.